♻️ move away from top-level module
This change places all the Python code in a separate `python` directory instead of directly at the top-level.
This is close to a `src`-layout, which is a recommended practise in Python development to avoid common pitfalls (such as the interpreter treating the top-level module as a package instead of using the installed package).