Add fixedMinimalTimeStep and limitMinimalTimeStep to MixingModel
The function fixedMinimalTimeStep can be used to allow for larger
timesteps which might cause mixtures to overflow the end of a channel,
but since no channel can be skipped should still be correct.

The function limitMinimalTimeStep allows to make sure that a simulation
step stops at an exact time, e.g. to save the state at that time.