Fix missing virtual destructors for base classes
A virtual destructor is necessary to make sure that a delete on a
pointer to the base class (e.g. use a channel pointer to delete a
rectangular channel) will actually delete all members of the child
class.
The failure to destruct all members is undefined behavior since
potentially crucial destructors won't be called (although in most cases
it is "just" a memory leak).