Fix string-int concatenation in exception message
Unfortunately, this works since the integer will be interpreted as a
character value, but will most likely result in an undesired result
(especially if the integer value is 0).
This can be simply fixed by adding corresponding calls to
std::to_string.
This kind of issue is reported by the compiler as a warning.